requirements.txt
requests==2.25.1
lxml==4.7.1
PyYAML~=5.4.1
alibabacloud_tea_openapi==0.3.1
alibabacloud_alidns20150109==2.0.2